Add files via upload
This program is a student grade evaluation system written in Python. It follows an object-oriented programming (OOP) approach and uses multiple classes to organize the functionality. The  code provides a simple, scalable structure to extend functionalities like storing data in a database, adding more grading criteria, or creating a GUI interface in the future.